requirements.txt
-e .
invoke==1.5.0
pytest==6.1.2
pytest-mock==3.3.1
pytest-parallel==0.1.0
httplib2==0.18.1
pydantic==1.7.2
requests==2.25.0
suds-community==0.8.5